requirements.txt
openai>=1.40.0
uagents==0.23.7
uagents-core==0.4.4
python-dotenv>=1.0.1